@@ -1,10 +1,10 @@
 [package]
 name = "foundry-fork-db"
 description = "Fork database used by Foundry"
-version = "0.15.0"
+version = "0.15.1"
 edition = "2021"
 # Remember to update clippy.toml as well
-rust-version = "1.83"
+rust-version = "1.85"
 authors = ["Foundry Contributors"]
 license = "MIT OR Apache-2.0"
 homepage = "https://github.com/foundry-rs/foundry-fork-db"
