@@ -1107,13 +1107,11 @@ mod tests {
                     Some(acc) => {
                         assert_eq!(
                             acc.nonce, new_acc.nonce,
-                            "The nonce was not changed in instance of index {}",
-                            idx
+                            "The nonce was not changed in instance of index {idx}"
                         );
                         assert_eq!(
                             acc.balance, new_acc.balance,
-                            "The balance was not changed in instance of index {}",
-                            idx
+                            "The balance was not changed in instance of index {idx}"
                         );
 
                         // comparing with db
@@ -1124,13 +1122,11 @@ mod tests {
 
                         assert_eq!(
                             db_address.nonce, new_acc.nonce,
-                            "The nonce was not changed in instance of index {}",
-                            idx
+                            "The nonce was not changed in instance of index {idx}"
                         );
                         assert_eq!(
                             db_address.balance, new_acc.balance,
-                            "The balance was not changed in instance of index {}",
-                            idx
+                            "The balance was not changed in instance of index {idx}"
                         );
                     }
                     None => panic!("Account not found"),
@@ -1176,7 +1172,7 @@ mod tests {
                             Ok(stg_db) => {
                                 assert_eq!(
                                     stg_db, *value,
-                                    "Storage in slot number {} in address {} do not have the same value", index, address
+                                    "Storage in slot number {index} in address {address} do not have the same value"
                                 );
 
                                 let db_result = {
@@ -1187,12 +1183,12 @@ mod tests {
 
                                 assert_eq!(
                                     stg_db, db_result,
-                                    "Storage in slot number {} in address {} do not have the same value", index, address
+                                    "Storage in slot number {index} in address {address} do not have the same value"
                                 )
                             }
 
                             Err(err) => {
-                                panic!("There was a database error: {}", err)
+                                panic!("There was a database error: {err}")
                             }
                         }
                     }
@@ -1237,8 +1233,7 @@ mod tests {
                         assert_eq!(
                             hash,
                             *block_hash_data.get(&key).unwrap(),
-                            "The hash in block {} did not match",
-                            key
+                            "The hash in block {key} did not match"
                         );
 
                         let db_result = {
@@ -1246,9 +1241,9 @@ mod tests {
                             *hashes.get(&key).unwrap()
                         };
 
-                        assert_eq!(hash, db_result, "The hash in block {} did not match", key);
+                        assert_eq!(hash, db_result, "The hash in block {key} did not match");
                     }
-                    Err(err) => panic!("Hash not found, error: {}", err),
+                    Err(err) => panic!("Hash not found, error: {err}"),
                 }
             }
         });
@@ -1315,7 +1310,7 @@ mod tests {
                             Ok(stg_db) => {
                                 assert_eq!(
                                     stg_db, *value,
-                                    "Storage in slot number {} in address {} doesn't have the same value", index, address
+                                    "Storage in slot number {index} in address {address} doesn't have the same value"
                                 );
 
                                 let db_result = {
@@ -1326,12 +1321,12 @@ mod tests {
 
                                 assert_eq!(
                                     stg_db, db_result,
-                                    "Storage in slot number {} in address {} doesn't have the same value", index, address
+                                    "Storage in slot number {index} in address {address} doesn't have the same value"
                                 );
                             }
 
                             Err(err) => {
-                                panic!("There was a database error: {}", err)
+                                panic!("There was a database error: {err}")
                             }
                         }
                     }
@@ -1374,8 +1369,7 @@ mod tests {
 
                         assert_eq!(
                             result_storage, *value,
-                            "Storage in slot number {} in address {} doesn't have the same value",
-                            index, address
+                            "Storage in slot number {index} in address {address} doesn't have the same value"
                         );
                     }
                 }
